feat: email rdv are now following a same selection flow
